Moved `ViewTransformed` methods to `Context`
See https://github.com/PistonDevelopers/graphics/issues/884
